requirements.txt
flake8==7.0.0
autopep8==2.1.0
pylint==3.1.0

django==5.0.3
python-decouple==3.8
psycopg2-binary==2.9.9

pytest-playwright==0.4.4
pixelmatch==0.3.0
